Stop resolving Uris encoded in the build protocol (#1607)
This makes a `hook/{build,link}.dart` no longer resolve uris encoded in the
`config.json`. It still validates that the uris that are encoded exist
when it decodes the json.

The creator of `config.json` has to ensure that the uris it encodes in
that config are existing file urls that the `hook/{build,link}.dart`
scripts can open as-is (i.e. without resolving against any base uri).

In many cases that means that an invoker should encode absolute uris in
there - or if it encodes relative uris - guarantees that they are
relative to the working directory the `hook/{build,link}.dart` is
invoked.

Issue https://github.com/dart-lang/native/issues/1605