Reset to path dependencies, make -wip versions & update changelog
We have published new versions of

  * `package:native_assets_builder`
  * `package:native_assets_cli`
  * `package:native_toolchain_c`

So we can switch the repository back to "development mode" and use path
dependencies.

(For consistency across packages we also add `publish_to: none` to the
`package:native_assets_cli` - as we have this in the other 2. Serving as
a reminder not to publish -wip versions)