[native_assets_builder] Don't cache hook invocations if env vars change (#1759)
This PR makes the native assets builder save the `Platform.environment` a hook is run in. Subsequent invocations check if the environment didn't change.

This removes the `includeParentEnvironment` parameter everywhere. It was always set to true in `dartdev` and `flutter_tools` (This will require a manual roll into the Dart SDK and Flutter tools.)

A follow up PR should restrict the list of environment variables (https://github.com/dart-lang/native/pull/1764), this PR is about caching correctness.

Bug: 

* https://github.com/dart-lang/native/issues/32