[native_assets_cli] Rename `supportedAssetTypes` to `buildAssetTypes` (#1786)
This PR changes the `supportedAssetTypes` to `buildAssetTypes`.

Hooks should not fail early if an asset type is not supported. This enables the following situations:

1. If some part of the Dart code requires an asset, but is not invoked transitively, the Dart code can still run in JIT, instead of failing on the build hook.
2. If an asset is not available, Dart code could fallback on a Dart implementation.
3. Hot reload hook invocations can omit the asset types that are not supported for hot reload (https://github.com/dart-lang/native/issues/1207#issuecomment-2505879943).

The PR changes both the Dart API and the JSON protocol. The JSON protocol passes and parses the old key for the time being to allow for a graceful transition until this rolls into a next Dart stable release and we can bump the lower SDK bound.

Side note:

* The refactorings somehow started disallowing version skew where an older SDK is used with a newer `package:native_assets_cli`, but we do allow this, we can fill in the blanks (see `pkgs/native_assets_cli/lib/src/model/build_config_CHANGELOG.md`). So this PR adds a lowest-parsable-version.