[native_assets_builder] Report dart sources as dependencies of hooks (#1780)
Closes: https://github.com/dart-lang/native/issues/1770

This PR adds the Dart sources to the dependencies in `HookResult`.

It also fixes the dep-file parsing w.r.t. to escapes.

### Implementation details

We're passing around `HookOutput`s which is the deserialized `output.json`. We could add the Dart sources as dependencies to it _after_ deserializing, but then the correspondence to the json is lost. So instead I've added an extra return value to the places where we pass `HookOutput` around.