[native_assets_cli] Move `BuildMode` to [native_toolchain_c] (#1800)
The first in a series of refactorings to address:

* https://github.com/dart-lang/native/issues/1738

We align `BuildMode` with `OptimizationLevel`, it's configured by the hook author in the build hook. Likely, every package should ship with max optimization level and build mode set to release. Only while developing a package (when it is the root package or path-depsed in) would the build mode be set to debug.

This will require a manual roll into dartdev and flutter tools due to the removal of `BuildMode` from the `native_assets_builder` API.

The native assets builder will still emit the field in the JSON, until we can bump the `native_assets_cli` SDK constraint to 3.7 stable. Then only people with older versions of the `native_assets_cli` package but a newer SDK break.