[native_assets_cli] Validate that paths are absolute in `Input` and `Output` (#1935)
Closes: https://github.com/dart-lang/native/issues/1605

Reuses the same validation (and error messages) for checking that files/directories exist and that their paths are absolute.

Also, adds a script to automatically update the `manifest.yaml` files. (It seems that `sort lines ascending` in vscode orders `/` and `_` differently than the string compiler in Dart. Hence the reordering.)