[native_assets_builder] Refactor API to `ProtocolExtension`s (#2089)
This PR introduces a concept of protocol extensions, which package together everything belonging to a protocol extension.

Closes: https://github.com/dart-lang/native/issues/2088

As a bonus, this better abstraction reduces LOC.

The `validation` methods are now all internal to the extensions. The `setup` methods stay exported because they are used extensively in hook-helper-packages (e.g. `native_toolchain_c`).