[native_assets_cli] Add paths to syntax views (#2107)
Bug: https://github.com/dart-lang/native/issues/1826

This PR adds the path that was navigated through the JSON to the syntax classes. This makes syntax errors more descriptive: `Unexpected value '123' for 'target_os'. Expected a String.` -> `Unexpected value '123' (int) for 'config.code.target_os'. Expected a String.`.

Because the data asset and code asset extensions don't know the internals of the base protocol extension, the extension points now expose both a `json` and a `jsonPath`. That path is then used to instantiate the syntax view of the extension, so it can give more precise error messages.

* For `CodeConfig` this comes naturally, it's a clean view.
* For `EncodedAsset` this is somewhat weird because it is half-used as value class: https://github.com/dart-lang/native/issues/2045

(The next PR will add a full syntax validation pass, at which point we should be able to address https://github.com/dart-lang/native/issues/2039.)