[native_assets_cli] Make `CodeAsset` `os` and `architecture` optional in JSON (#2139)
Bug: https://github.com/dart-lang/native/pull/2138

This makes the OS and architecture optional in the syntax and the internals in the semantic API (after https://github.com/dart-lang/native/pull/2138 they are no longer available in the semantic API).

The validation logic still checks that the OS and architecture is consistent with the `CodeConfig`.

(We cannot remove them yet, because link hooks can rely on these fields being present in the input if they were present in the build hook output. The previous PR fixed one such case. So we'll leave that cleanup to later once the removal of the API has rolled everywhere.)