[native_assets_cli] Add base path for user-defines (#2209)
Bug: https://github.com/dart-lang/native/issues/39
Closes: https://github.com/dart-lang/native/issues/2187
Closes: https://github.com/dart-lang/native/issues/2201

### Relative paths in user-defines accessible as absolute paths in hooks

This PR enables resolving relative paths in user-defines as absolute paths. The `input.json` now is structured as follows:

```json
{
  "user_defines": {
    "workspace_pubspec": {
      "base_path": "/private/var/folders/2y/mngq9h194yzglt4kzttzfq6800klzg/T/QqYcWN/user_defines/pubspec.yaml",
      "defines": {
        "user_define_key": "user_define_value",
        "user_define_key2": {
          "foo": "bar"
        },
        "some_file": "assets/data.json"
      }
    },
  },
}
```

We don't want to expose the workspace/pubspec path (except for resolving user-defines), so this can be done in the Dart API as follows:

```dart
extension type HookInputUserDefines._(HookInput _input) {
  /// The value for the user-define for [key] for this package.
  ///
  /// This can be arbitrary JSON/YAML if provided from the SDK from such source.
  /// If it's provided from command-line arguments, it's likely a string.
  Object? operator [](String key) => // already present before this PR

  /// The absolute path for user-defines for [key] for this package.key
  ///
  /// The relative path passed as user-define is resolved against the base path.
  /// For user-defines originating from a JSON/YAML, the base path is this
  /// JSON/YAML. For user-defines originating from command-line aruments, the
  /// base path is the working directory of the command-line invocation.
  ///
  /// If the user-define is `null` or not a [String], returns `null`.
  Uri? path(String key) => // added in this PR.
}
```

This PR does _not_ yet add support for command-line user-defines. But the JSON has been set up so that it's extensible.

### Move reading user-defines from flutter_tools/dartdev to `NativeAssetsBuildRunner`

This PR changes the responsibility to read the user-defines from the pubspec to the `NativeAssetsBuildRunner`.

1. This ensures all SDKs have the same logic for reading user-defines
2. This ensures all SDKs have the same caching behavior for files containing user-defines changing on disk.

This PR adds the pubspec.yaml in the `HookOutput.dependencies` to ensure the build runner is reinvoked if the pubspec changes.

The API for the `NativeAssetsBuildRunner` changes to take the `pubspec.yaml` uri rather than user-defines. Consequently, this PR will need a manual roll into dartdev and flutter_tools.