[native_assets_cli] Curate public toplevel functions (#2249)
This PR tries the curate the toplevel functions which are in the public API of what will be `package:hooks`,   `package:code_assets` and `package:data_assets`.

The validate functions for the base protocol are moved into a `ProtocolBase` class. (These need to be public because `package:native_assets_builder` uses them. And these need to be in `package:native_assets_cli` because the `test` methods use them.)

The Dart docs can be inspected by `pkgs/native_assets_cli$ dart doc .`.

TODOs:

* This PR does not yet curate extension types, and extensions.

Context:

* Work before splitting up the package: https://github.com/dart-lang/native/issues/999