[json_syntax_generator] Make `JsonReader` private (#2446)
Addressing https://github.com/dart-lang/native/pull/2445#discussion_r2230432015

We can't make `JsonObjectSyntax` private, we use it as the type for `Map<String, Object?>` to enable overriding fields with this type with a class extending `JsonObjectSyntax` in subclasses.

```dart
AssetSyntax({required JsonObjectSyntax? encoding, required String type})
```

And then `DataAssetEncodingSyntax` is a valid type to be passed to `encoding`

```dart
class DataAssetEncodingSyntax extends _JsonObjectSyntax {
```

It would be cleaner if the runtime library would be a standalone package, and all the generated files would import that. But that requires us to publish that library on pub.dev for rolling hooks_runner into flutter_tools. So maybe the simplest solution is to keep it as is for now. (If we ever decide to make the syntax generator a fully supported project we can do it the clean way.) WDYT?