[pub_formats] Make nullable params optional (#2459)
For the generated syntax inside `package:hooks`, `code_assets`, and `data_assets` all nullable parameters are `required`. This is to help us ensure that we don't forget a param in the semantic API. (The generated syntax is wrapped with an API handcrafted to be nice for our users.)

However, in `package:pub_formats` we don't have a semantic API wrapping the syntax, we directly use the syntax classes in `dartdev` and friends. In such case it's unwanted to have to provide every param. (For example when constructing a `pubspec.yaml` we want to omit all the optional sections.)

So, we add a param to the generator to control this.