[native_toolchain_c] Fix page size assertions in tests (#2838)
Fixes https://github.com/dart-lang/native/issues/2837

The tests that asserted on the page size started failing on macOS (see https://github.com/dart-lang/native/issues/2837). According to Gemini they weren't asserting on the right thing: The virtual memory address within a section (this is what the test was asserting) is not the same thing as the alignment of the ELF load segment (this is what the test supposedly wanted to test to verify that the page size setting was honored).